backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-filter==23.5
drf-yasg==1.21.7
celery==5.3.4
redis==5.0.1
python-dotenv==1.0.0
PyYAML==6.0.1
psycopg2-binary==2.9.9